requirements/requirement.txt
tk
customtkinter
Pillow
requests
cv2
matplotlib
tensorflow